refactors view my team view
