+++ src/components/player/profile-modal-components/PlayerProfileBanner.tsx
@@ -1,17 +1,20 @@
 import { useEffect, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { CardTier, IProAthlete } from '../../../types/athletes';
+import { CardTier } from '../../../types/athletes';
 import { ScrummyDarkModeLogo } from '../../branding/scrummy_logo';
 import { CircleDollarSign } from 'lucide-react';
+import { usePlayerData } from '../provider/PlayerDataProvider';
 
 interface Props {
-  player: IProAthlete;
 }
 
-export default function PlayerProfileBanner({ player }: Props) {
+export default function PlayerProfileBanner({ }: Props) {
+
+  const { player } = usePlayerData();
+
   // Fallback chain like PlayerGameCard
-  const primaryImageUrl = player.image_url ?? null;
-  const teamFallbackUrl = player.team_id
+  const primaryImageUrl = player?.image_url ?? null;
+  const teamFallbackUrl = player?.team_id
     ? `https://athstat-landing-assets-migrated.s3.us-east-1.amazonaws.com/logos/${player.team_id}-ph-removebg-preview.png`
     : null;
   const [src, setSrc] = useState<string | null>(primaryImageUrl ?? teamFallbackUrl);
@@ -32,9 +35,11 @@ export default function PlayerProfileBanner({ player }: Props) {
     }
   };
 
-  const pr = player.power_rank_rating ?? 0;
+  const pr = player?.power_rank_rating ?? 0;
   const cardTier: CardTier =
     pr <= 69 ? 'bronze' : pr >= 70 && pr < 80 ? 'silver' : pr >= 90 ? 'blue' : 'gold';
+
+  if (!player) return;
 
   return (
     <div
@@ -43,16 +48,16 @@ export default function PlayerProfileBanner({ player }: Props) {
         cardTier === 'gold' && 'bg-gradient-to-br from-yellow-300 via-yellow-500 to-yellow-700 ',
         cardTier === 'silver' && 'bg-gradient-to-br from-gray-300 via-gray-400 to-gray-600',
         cardTier === 'bronze' &&
-          'bg-gradient-to-br from-amber-600 via-amber-800 to-amber-900 text-white',
+        'bg-gradient-to-br from-amber-600 via-amber-800 to-amber-900 text-white',
         cardTier === 'blue' &&
-          'bg-gradient-to-br from-purple-600 via-blue-800 to-purple-900 text-white',
+        'bg-gradient-to-br from-purple-600 via-blue-800 to-purple-900 text-white',
         cardTier === 'blue' && 'animate-glow'
       )}
     >
-      {player.price !== undefined && (
+      {player?.price !== undefined && (
         <div className="absolute top-2 left-2 z-40 flex flex-row items-center gap-1 bg-black/10 text-white rounded-md px-2 py-1">
           <CircleDollarSign className="w-4 h-4" />
-          <p className="text-sm font-bold">{player.price}</p>
+          <p className="text-sm font-bold">{player?.price}</p>
         </div>
       )}
       {src ? (
